Correcciones

+++ practica2.3/ejercicio13.c
@@ -3,35 +3,49 @@
 #include <unistd.h>
 #include <stdlib.h>
 
-int main(int argc, char *argv[]) {
-    printf("El PID es: %i\n", getpid());
+volatile int borra = 1;
+volatile int alarma = 0;
 
-    sigset_t set;
-    sigemptyset(&set);
-    sigaddset(&set, SIGUSR1);
-    sigprocmask(SIG_BLOCK, &set, NULL);
-    
-    int segs = atoi(argv[1]);
-    sleep(segs);
+void manejador(int signal){
+	if(signal == SIGALRM) alarma = 1;
+	if(signal == SIGUSR1) borra = 0;
+}
 
-    sigset_t pendset;
-    sigemptyset(&pendset);
+int main(int argc, char *argv[]){
+	printf("Mi PID es: %i\n", getpid());
+
+	struct sigaction act1;
+	struct sigaction act2;
 	
-    if(sigpending(&pendset) == -1){
-	perror("Error sispending");
-	return -1;
-    }
+	if(sigaction(SIGALRM, NULL, &act1) == -1 || sigaction(SIGUSR1, NULL, &act2) == -1){
+		perror("Error sigaction");
+		return -1;
+	}
 
-    switch(sigismember(&pendset, SIGUSR1)){
-        case -1:
-            perror("Error sigismember");
-            return -1;
-        break;
-        case 0:
-            printf("Borrado");
-            unlink(argv[0]);
-        break;
-    }
+	act1.sa_handler = manejador;
+	act2.sa_handler = manejador;
 
-    return 0;
+	int sa1 = sigaction(SIGALRM, &act1, NULL);
+	int sa2 = sigaction(SIGUSR1, &act2, NULL);
+
+	if(sa1 == -1 || sa2 == -1){
+		perror("Error sigaction");
+		return -1;
+	}
+
+	sigset_t set;
+	sigfillset(&set);
+	sigdelset(&set, SIGALRM);
+	sigdelset(&set, SIGUSR1);
+
+	alarm(atoi(argv[1]));
+
+	sigsuspend(&set);
+
+	if(borra == 1 && alarma == 1){
+		printf("Borrado");
+		unlink(argv[0]);
+	}
+
+	return 0;
 }
